+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/RNObject.kt
@@ -2,7 +2,6 @@
 package host.exp.exponent
 
 import host.exp.exponent.analytics.EXL
-import host.exp.expoview.BuildConfig
 import java.lang.reflect.Constructor
 import java.lang.reflect.Field
 import java.lang.reflect.InvocationTargetException
@@ -36,7 +35,7 @@ class RNObject {
   // required for "unversioned" flavor check
   fun loadVersion(version: String): RNObject {
     try {
-      clazz = if (version == UNVERSIONED || BuildConfig.FLAVOR == "unversioned") {
+      clazz = if (version == UNVERSIONED) {
         if (className.startsWith("host.exp.exponent")) {
           Class.forName("versioned.$className")
         } else {