+++ apps/expo-go/android/expoview/src/main/java/versioned/host/exp/exponent/modules/universal/ConstantsBinding.kt
@@ -27,14 +27,14 @@ class ConstantsBinding(
       this["nativeAppVersion"] = ExpoViewKernel.instance.versionName
       this["nativeBuildVersion"] = Constants.ANDROID_VERSION_CODE
       this["supportedExpoSdks"] = Constants.SDK_VERSIONS_LIST
-      this["appOwnership"] = appOwnership
+      this["appOwnership"] = "expo"
       this["executionEnvironment"] = executionEnvironment.string
 
       this.putAll(experienceProperties)
 
       this["platform"] = mapOf(
         "android" to mapOf(
-          "versionCode" to if (appOwnership == "expo") null else Constants.ANDROID_VERSION_CODE
+          "versionCode" to null
         )
       )
       this["isDetached"] = false
@@ -49,17 +49,8 @@ class ConstantsBinding(
     }
   }
 
-  override fun getAppOwnership(): String {
-    return "expo"
-  }
-
   private val executionEnvironment: ExecutionEnvironment
     get() = ExecutionEnvironment.STORE_CLIENT
-
-  override fun getOrCreateInstallationId(): String {
-    // Override scoped installationId from ConstantsService with unscoped
-    return exponentSharedPreferences.getOrCreateUUID()
-  }
 
   companion object {
     private fun convertPixelsToDp(px: Float, context: Context): Int {