+++ packages/@expo/cli/src/start/server/metro/createServerRouteMiddleware.ts
@@ -4,17 +4,20 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { ExpoResponse } from '@expo/server';
-import { createRequestHandler } from '@expo/server/build/vendor/http';
+
+import type { ProjectConfig } from '@expo/config';
 import requireString from 'require-from-string';
 import resolve from 'resolve';
+import resolveFrom from 'resolve-from';
 import { promisify } from 'util';
 
 import { ForwardHtmlError } from './MetroBundlerDevServer';
 import { bundleApiRoute } from './bundleApiRoutes';
 import { fetchManifest } from './fetchRouterManifest';
 import { getErrorOverlayHtmlAsync, logMetroError, logMetroErrorAsync } from './metroErrorInterface';
+import { warnInvalidWebOutput } from './router';
 import { Log } from '../../../log';
+import { CommandError } from '../../../utils/errors';
 
 const debug = require('debug')('expo:start:server:metro') as typeof console.log;
 
@@ -33,8 +36,19 @@ export function createRouteHandlerMiddleware(
     baseUrl: string;
     getWebBundleUrl: () => string;
     getStaticPageAsync: (pathname: string) => Promise<{ content: string }>;
+    config: ProjectConfig;
   }
 ) {
+  if (!resolveFrom.silent(projectRoot, 'expo-router')) {
+    throw new CommandError(
+      'static and server rendering requires the expo-router package to be installed in your project.'
+    );
+  }
+
+  const { ExpoResponse } = require('@expo/server') as typeof import('@expo/server');
+  const { createRequestHandler } =
+    require('@expo/server/build/vendor/http') as typeof import('@expo/server/build/vendor/http');
+
   return createRequestHandler(
     { build: '' },
     {
@@ -89,6 +103,7 @@ export function createRouteHandlerMiddleware(
               }
             );
           } catch (staticError: any) {
+            debug('Failed to render static error overlay:', staticError);
             // Fallback error for when Expo Router is misconfigured in the project.
             return new ExpoResponse(
               '<span><h3>Internal Error:</h3><b>Project is not setup correctly for static rendering (check terminal for more info):</b><br/>' +
@@ -110,6 +125,11 @@ export function createRouteHandlerMiddleware(
         logMetroError(projectRoot, { error });
       },
       async getApiRoute(route) {
+        const { exp } = options.config;
+        if (exp.web?.output !== 'server') {
+          warnInvalidWebOutput();
+        }
+
         const resolvedFunctionPath = await resolveAsync(route.page, {
           extensions: ['.js', '.jsx', '.ts', '.tsx'],
           basedir: options.appDir,
@@ -127,7 +147,7 @@ export function createRouteHandlerMiddleware(
 
         try {
           debug(`Bundling middleware at: ${resolvedFunctionPath}`);
-          return requireString(middlewareContents);
+          return requireString(middlewareContents.src, middlewareContents.filename);
         } catch (error: any) {
           if (error instanceof Error) {
             await logMetroErrorAsync({ projectRoot, error });