+++ packages/expo-constants/android/src/main/java/expo/modules/constants/ConstantsService.kt
@@ -3,14 +3,11 @@ package expo.modules.constants
 import org.apache.commons.io.IOUtils
 
 import expo.modules.core.interfaces.InternalModule
-import expo.modules.core.utilities.EmulatorUtilities
 import expo.modules.interfaces.constants.ConstantsInterface
 
 import android.os.Build
 import android.util.Log
 import android.content.Context
-import android.content.pm.PackageInfo
-import android.content.pm.PackageManager
 
 import java.io.FileNotFoundException
 import java.lang.Exception
@@ -28,7 +25,6 @@ open class ConstantsService(private val context: Context) : InternalModule, Cons
     ?: 0
 
   private val sessionId = UUID.randomUUID().toString()
-  private val exponentInstallationId: ExponentInstallationId = ExponentInstallationId(context)
 
   enum class ExecutionEnvironment(val string: String) {
     BARE("bare"),
@@ -39,46 +35,26 @@ open class ConstantsService(private val context: Context) : InternalModule, Cons
   override fun getExportedInterfaces(): List<Class<*>> = listOf(ConstantsInterface::class.java)
 
   override fun getConstants(): Map<String, Any?> {
-    val constants = mutableMapOf(
+    return mutableMapOf(
       "sessionId" to sessionId,
       "executionEnvironment" to ExecutionEnvironment.BARE.string,
       "statusBarHeight" to statusBarHeightInternal,
       "deviceName" to deviceName,
-      "isDevice" to isDevice,
       "systemFonts" to systemFonts,
       "systemVersion" to systemVersion,
-      "installationId" to getOrCreateInstallationId(),
       "manifest" to appConfig,
-      "platform" to mapOf("android" to emptyMap<String, Any>())
+      "platform" to mapOf<String, Map<String, Any>>("android" to emptyMap())
     )
-
-    try {
-      val pInfo = context.packageManager.getPackageInfo(context.packageName, 0)
-      constants["nativeAppVersion"] = pInfo.versionName
-
-      val versionCode = getLongVersionCode(pInfo).toInt()
-      constants["nativeBuildVersion"] = versionCode.toString()
-    } catch (e: PackageManager.NameNotFoundException) {
-      Log.e(TAG, "Exception: ", e)
-    }
-
-    return constants
   }
 
   // Just use package name in vanilla React Native apps.
   override fun getAppScopeKey(): String? = context.packageName
 
-  override fun getAppOwnership() = "guest"
-
   override fun getDeviceName(): String = Build.MODEL
-
-  override fun getIsDevice() = !EmulatorUtilities.isRunningOnEmulator()
 
   override fun getStatusBarHeight() = statusBarHeightInternal
 
   override fun getSystemVersion(): String = Build.VERSION.RELEASE
-
-  open fun getOrCreateInstallationId(): String = exponentInstallationId.getOrCreateUUID()
 
   // From https://github.com/dabit3/react-native-fonts
   override fun getSystemFonts() = listOf(
@@ -116,15 +92,5 @@ open class ConstantsService(private val context: Context) : InternalModule, Cons
       val dp = px / (metrics.densityDpi / 160f)
       return dp.toInt()
     }
-
-    private val isRunningOnEmulator: Boolean
-      get() = EmulatorUtilities.isRunningOnEmulator()
-
-    private fun getLongVersionCode(info: PackageInfo) =
-      if (Build.VERSION.SDK_INT >= 28) {
-        info.longVersionCode
-      } else {
-        info.versionCode.toLong()
-      }
   }
 }