+++ packages/expo-updates/android/src/androidTest/java/expo/modules/updates/procedures/StateMachineSerialExecutorQueueTest.kt
@@ -3,6 +3,8 @@ package expo.modules.updates.procedures
 import android.os.Handler
 import android.os.HandlerThread
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
+import androidx.test.platform.app.InstrumentationRegistry
+import expo.modules.updates.logging.UpdatesLogger
 import expo.modules.updates.statemachine.UpdatesStateEvent
 import expo.modules.updates.statemachine.UpdatesStateValue
 import org.junit.Test
@@ -16,7 +18,12 @@ class StateMachineSerialExecutorQueueTest {
   fun test_SerialExecution() {
     val latch = CountDownLatch(3)
 
+    val context = InstrumentationRegistry.getInstrumentation().targetContext
+    val logger = UpdatesLogger(context)
+
     class TestStateMachineProcedure : StateMachineProcedure() {
+      override val loggerTimerLabel = "timer-test"
+
       var executionTime: Long = 0
 
       override fun run(procedureContext: ProcedureContext) {
@@ -36,18 +43,21 @@ class StateMachineSerialExecutorQueueTest {
       }
     }
 
-    val executorQueue = StateMachineSerialExecutorQueue(object : StateMachineProcedure.StateMachineProcedureContext {
-      override fun processStateEvent(event: UpdatesStateEvent) {
-      }
+    val executorQueue = StateMachineSerialExecutorQueue(
+      logger,
+      object : StateMachineProcedure.StateMachineProcedureContext {
+        override fun processStateEvent(event: UpdatesStateEvent) {
+        }
 
-      @Deprecated("Avoid needing to access current state to know how to transition to next state")
-      override fun getCurrentState(): UpdatesStateValue {
-        return UpdatesStateValue.Idle
-      }
+        @Deprecated("Avoid needing to access current state to know how to transition to next state")
+        override fun getCurrentState(): UpdatesStateValue {
+          return UpdatesStateValue.Idle
+        }
 
-      override fun resetState() {
+        override fun resetState() {
+        }
       }
-    })
+    )
 
     val procedure1 = TestStateMachineProcedure()
     val procedure2 = TestStateMachineProcedure()