+++ packages/@expo/metro-runtime/CHANGELOG.md
@@ -6,9 +6,17 @@
 
 ### 🎉 New features
 
+- Add error handling for Metro build errors in split chunks. ([#26609](https://github.com/expo/expo/pull/26609) by [@EvanBacon](https://github.com/EvanBacon))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+## 3.1.1 - 2024-01-18
+
+### 🐛 Bug fixes
+
+- Fixed lazy component errors on Android. ([#26464](https://github.com/expo/expo/pull/26464) by [@kudo](https://github.com/kudo))
 
 ## 3.1.0 — 2023-12-12