+++ packages/expo-manifests/CHANGELOG.md
@@ -15,6 +15,14 @@
 - [Android] Remove unsafe internal mutation capability. ([#26229](https://github.com/expo/expo/pull/26229) by [@wschurman](https://github.com/wschurman))
 - Rename manifest classes. ([#26234](https://github.com/expo/expo/pull/26234), [#26235](https://github.com/expo/expo/pull/26235), [#26257](https://github.com/expo/expo/pull/26257) by [@wschurman](https://github.com/wschurman))
 
+## 0.13.2 - 2024-01-18
+
+_This version does not introduce any user-facing changes._
+
+## 0.13.1 - 2024-01-10
+
+_This version does not introduce any user-facing changes._
+
 ## 0.13.0 — 2023-12-12
 
 _This version does not introduce any user-facing changes._